+++ datumaro/plugins/voc_format/extractor.py
@@ -12,7 +12,8 @@ from datumaro.components.annotation import (
     AnnotationType, Bbox, CompiledMask, Label, Mask,
 )
 from datumaro.components.extractor import DatasetItem, SourceExtractor
-from datumaro.util.image import Image, find_images
+from datumaro.components.media import Image
+from datumaro.util.image import find_images
 from datumaro.util.mask_tools import invert_colormap, lazy_mask
 from datumaro.util.meta_file_util import is_meta_file, parse_meta_file